pyproject.toml
[tool.poetry]
name = "knobby"
version = "0.1.0"
description = "Генерация ассетов для вращающихся кнобов (поворотных ручек на устройствах, например гитарных педалях)"
authors = ["The <at.thealchemist@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^21.2.0"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.11b1"
isort = "^5.10.1"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
